First batch of fixes
