+++ .travis.yml
@@ -10,6 +10,7 @@ service:
 
 env:
   - GO111MODULE=on GOPROXY=https://proxy.golang.org
+  - REG_TOKEN=$GITHUB_TOKEN
 
 matrix:
   fast_finish: true
